done all services, changed UserScore to float, finished seeders, finished mapper + more

+++ Project/Models/DTOs/MovieDTOs/MovieUpdateDTO.cs
@@ -3,7 +3,7 @@
     public class MovieUpdateDTO
     {
         public Guid Id { get; set; }
-        public int? UserScore { get; set; }
+        public float? UserScore { get; set; }
 
     
     }
