+++ Project/Models/DTOs/UserMovie(Review)DTOs/UserMovie(Review)FormattedDTO.cs
@@ -5,7 +5,7 @@
         public string UserName { get; set; }
         public string MovieTitle { get; set; }
         public string Review { get; set; }
-        public int UserScore { get; set; }
+        public float UserScore { get; set; }
         public int Duration { get; set; }
     }
 }
